exercises for part 4

+++ bloglist-backend/controllers/blogs.js
@@ -1,15 +1,25 @@
+const jwt = require("jsonwebtoken");
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
+const User = require("../models/user");
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   response.json(blogs);
 });
 
 blogsRouter.post("/", async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
+  const user = await User.findById(decodedToken.id);
+
   const blog = new Blog({
     ...request.body,
     likes: request.body.likes === undefined ? 0 : request.body.likes,
+    user: user._id,
   });
 
   if (!blog.title || !blog.url) {
@@ -17,10 +27,29 @@ blogsRouter.post("/", async (request, response) => {
   }
 
   const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
   response.status(201).json(savedBlog);
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
+  const blog = await Blog.findById(request.params.id);
+  if (!blog) {
+    return response.status(404).end();
+  }
+
+  if (blog.user.toString() !== decodedToken.id.toString()) {
+    return response
+      .status(401)
+      .json({ error: "only the creator can delete this blog" });
+  }
+
   await Blog.findByIdAndDelete(request.params.id);
   response.status(204).end();
 });
